Removing committee members
Saving form data from a tab sends only the data from that tab.  This was
causing a problem when the student deletes all committee members from
the form because they weren't properly being deleted from the cached
data.

This solution looks for the committee_chair_attributes key, which is a
required field, so it should be among the data keys if we are submitting
from the 'My Advisors' tab.  If we're on that tab, and we don't see a
key for the committee members, then we can assume that the student
deleted the members from the form, so we should delete them from the
cached data as well.