Cast `#*_embargoed` values to boolean in the model
Various parts of the codebase set these flags to `true`/`false` or
`"true"`/(`""`|`"false"|`nil`). These are interpreted in `SolrDocument`, but the
model itself is unable to tell a caller whether these values are embargoed with
any reliability.

The fix included here is to override the setters and getters for these values
to cast aggressively to a Boolean value. Since `SolrDocument` already has this
logic baked in, this ensures both the model and the index give the intended
value. We can deprecate setting with string values and remove the overrides once
users of this approach (notably the submission forms) have been cleaned up.

Fixes #1763