Address security issue CVE-2015-9284
Vulnerable versions: <= 1.9.0
Patched version: No fix
The request phase of the OmniAuth Ruby gem is vulnerable to Cross-Site Request Forgery when used as part of the Ruby on Rails framework, allowing accounts to be connected without user intent, user interaction, or feedback to the user. This permits a secondary account to be able to sign into the web application as the primary account.

More info available at https://nvd.nist.gov/vuln/detail/CVE-2015-9284

After review, laevigata does not appear to be particularly vulnerable to
this issue, as it would require that the Emory Shibboleth system also be
compromised. However, to be on the safe side, we are implementing the
recommended fix. Until the omniauth gem itself can be upgraded (long
thread about why this is taking so long available here:
https://github.com/omniauth/omniauth/pull/809) the recommendation is to
install the omniauth-rails_csrf_protection gem and ensure all
authentication requests happen via POST.

This PR installs the required gem, and all of our authentication
requests already use a POST protocol.