Fix a bug handling open access ETDs
Whether the student requests and embargo or not, the applicaton
places all new ETDs under an embargo to prevent them from being
visible in search results until the graduation job has been run.
This results in a situation where the student has requested the
ETD to be under embargo, but the visibility should display as open
access.  The new code adds a test for this condition and the
necessary code to handle the condition correctly.

Prior to this commit, the code assumed that if any embargo was present,
visibility should default to the lowest "Files Only" embargo
restriction level. This resulted in a number of open access ETDs being
published with erroneous "Restricted: files only" visibliity.