Handle edge cases in embargo requests and visiblity settings
There are a numbrer of edge cases that "shouldn't" be able to occur
in the application but appear to be possible based on an analysis
of production data.

When the user has not requested an embargo by selecting the
"None - open access immediately" option, we always set the visiblity
to open access and log an error when abstract, toc, or file embargo
flags are not set to false.

When the user has requested an embargo by selecting a non-zero embargo
length, but abstract, toc, or file embargo flags are not set, we return
visibility as restricted since we can't effectively interpret the user's
intent and log an error to track the inconsistent state.