Index `embargo_length` as a singluar, searchable, stored value in Solr
Previously `embargo_length` was indexed as a stored, but not indexed,
symbol (`_ssm`).  This prevents searching Solr for the attribute,
which has made debugging certain embargo and graduation bugs difficult.

The changes index the attribute as a single-value symbol (`_ssi`).
Since the student can only choose one value for the requested embargo
duration, this aligns the data model more closely with the problem domain.