Refactor the GraduationService for improved performance
The current GraduationService takes over 30 minutes to run against
production data.  Analysis indicated that the current query for
candiate ETD works takes over 10 minutes to process and returns
over 6000 ETDs that have already been processed by the service and
are already in a 'published' state.

This commit queries only for ETDs that are 'approved' which is the only
group that are eligible for graduation and publication.  Querying in
this way is much faster and also eliminates the need for additional
error handling code, thereby simplifying the overall service.

The commit also removes dead code for the GraduationdateService.