Remove unused workflow configuration from the application
Hyrax creates a permission template for every configured workflow
for every admin_set; regardless of whether the workflow is enabled
for the admin set.
https://github.com/samvera/hyrax/blob/v2.9.3/app/services/hyrax/workflow/workflow_importer.rb#L34

The default one-step unmediated deposit workflow generated when
installing Hyrax is not used anywhere in the ETD application.
This change removes the unneeded workflow file to speed up
testing  and deployment.