Update gem dependencies
Version 11.0.5 of hydra-access-controls was yanked from rubygems.org,
making it impossible to deploy the application or run the test suite in
new environments.  A newer version of hyrdra-access-controls was released
and a patch version of hyrax released to accomodate the new dependencies.

This change set results from running `bundle update hyrax`