Refactor GraduationJob test to be more resilient
The previous implementation of the job depended on starting with
Fedora in a clean state by using the `:clean` tag.  Erasing Fedora
is an expensive operation that we'd like to avoid when possible.

This change set captures the specific ID of the ETD created for the
test so that we don't have to rely on a clean Fedora environment and
assuming the `last` created ETD is the right one to test..