Refactor attribute "embargo_length" name for clarity
We've had a term named `embargo_length` which stores the embargo
duration a student is requesting be applied once they graduate.

This term reads very similar to `embargo_release_date` which is
the actual date an embargo will expire once a student has graduated.
Both of these fields frequently occur in close proximilty in the code
making it hard to remember which is the requestsed, but not final,
value and which is the actual value applied when the work is published.

This change just renames the requested value from
`embargo_length` to `requested_embargo_duration`
in order to help make it easier to distinquish role of the two values
in the code.