Speed up contact page test (#2177)
The previous test loaded and executed remote javascript from the
libwizard site.  While it's useful to know that this code is running
successfully and inserting the expected iframe form into the ETD Help
page, the test takes a very long time to run and the javascript does
not always load from the remote site before the test times out.

This change flags that test as a smoke_test that is not run during
regular runs of the test suite.  It also adds a view test that ensures
the default Hyrax contact form is replaced by the local form which
embeds the expected libwizard script.