Add a fix for the ETD publication bug (#2187)
The GraduationJob is leaving certain works in a state where they
cannot be viewed after graduation. Attempting to view the ETDs in question
gives the message:
>The work is not currently available because it has not yet completed the publishing process

This change adds a test that puts the ETD in the same pregraduation state:
1. The student initally submits the ETD with an embargo requested &
a placeholder pre-graduation-embargo is created
2. The student changes their mind and choosed to submit the ETD as
open access before they gradute, but the placeholder embargo does not
get removed
3. When the GraduationJob runs it leaves the ETD in an undiscoverable state
because of the left-over embargo conflicts with the request for open access

The previous code attempted to find an existing pre-graduation embargo and
update the expiration date and visibility.

The fix applied here deletes any embargo applied prior to the student's
graduation. This prevents "open access" ETDs from being published with orphan
embargoes left over from the setps described above. The graduation job applies
embargoes only when a valid embargo request is still in place at graduation.