Consistently index degree_awarded as a date
We need to use the value stored in `degree_awarded` to support
listing ETDs for indexing by Google Scholar.

Because the code sometimes treated the field as a string and sometimes
treated it as a date or time object, the data was not being consistently
indexed into solr.

The changes here ensure that the value stored in Fedora is always a
date type value and ensures that the serialization stored in Solr
is consistently sortable.