Add fallback to old Solr index style for degree_awarded
Going forward, we want to index degree_awarded as a singular date
(degree_awarded_dtsim), but our current index has it indexed as a
multi-valued field (degree_awarded_dtsim).

This patch allows the application to fall back to looking for the
older indexing style when we're running on environments that have
not had the full repo reindexed yet.