Rename "report.csv" to "etd_report.csv"
We have other reports in the system and might want to add others in
the future - so calling the new ETD Report just `report` might be
too generic.

This change renames references to report or nightly_report to etd_report
to help make things a little more explicit if we ever have to come
back to this code in the future.