Run each GraduationJob immediately (instead of queuing for later)
Stop running GraduationJobs in the background, since we're intending to either
1) Run the emory:graduation task from the console and want to watch the
ouput directly without burying it in a background job.
OR
2) Run the GraduationService from the UI where we would want to wrap the
whole process in a single job

Either way, there's no processing or performance benefit from processing
the GraduationJobs for individual ETDs via the background scheduler.

This change makes the whole service easier to monitor and debug.