Handle timezones for graduation dates
Once we have the graduation date, convert it to a Time class object
so it gets the appropriate server timezone applied - i.e. a graduation
date of "May 15,2020" gets treated as occurring at midnight Eastern
instead of UTC.

This is the preferred format when assigning `degree_awarded` to an
ETD and prevents the warning:
`Assigning degree_awarded= to a string is deprecated. Casting to a Date.`