Proquest Preliminary Refactor (#2265)
* Handle timezones for graduation dates

Once we have the graduation date, convert it to a Time class object
so it gets the appropriate server timezone applied - i.e. a graduation
date of "May 15,2020" gets treated as occurring at midnight Eastern
instead of UTC.

This is the preferred format when assigning `degree_awarded` to an
ETD and prevents the warning:
`Assigning degree_awarded= to a string is deprecated. Casting to a Date.`

* Refactor to remove instance variables from ProquestJob

* Get the ProquestJob test working

* Update the test environment to have a valid DERIVATIVES_PATH
(i.e. don't default to /opt/downloads)
* Inline Proquest completion since it doesn't serve any purpose to
run it as a delayed job (other than complicating testing)
* Fix the test setup for the ProquestJob so we can sucessfully
run the job
* Stub out characterize (so we don't have to install FITS for CI)
and provide fixture files with appropriate characterization output
* Use Date.current instead of Time.zone.today to make it clearer
that we're saving dates without time offsets
* Add a Laney PhD graduate to the registrar data to match the
etd setup in the job test

* Merge GraduationService tests into a single spec

Use contexts to scope which test need expesive setup rather than
splitting them into a separate test file.

* Move the .submit_to_proquest? logic

Move the business logic for whether to submit an ETD to ProQuest
out of the job and into the ProquestBehaviors so that it can be
called directly on the ETD in question.

Also stub out expensive Sipity setup to make tests run faster.

* Fix depositor creation in etd factory