Pass registrar data to jobs
When the GraduationService finds registrar data for a graduated ETD,
pass it to the the GraduationJob and ProquestJob.

TODO: refactor ProquestBehaviors to use the passed registrar data
instead of looking the data up from the registrar file on disk.

This allows us to delete the registrar data once the GraduationService
completes it's run.