Refactor tests to isolate test environment timezone (#2269)
ISSUE:
Depending on the system for the test suite run, the "local" ruby timezone
can either be something like
* UTC+5:00 "Eastern Time (US & Canada)"
* UTC+6:00 "America/Chicago"
* +0000     UTC
So, tests that check for a literal time may fail due to mismatches in
timezone offsets.  This occurs mostly between runs in local developer
environments (typically set to Eastern or Central time) and runs on
CircleCI (set to UTC).

CHANGE:
Update tests to either match just on the date portion of timestamps
or explicitly set timezones for the specific test.