Display e-mail on Manage Users view in admin dashboard
When we get bounce notifications from AWS, the only user data we
have available is the e-mail address.  Since this is typically the
post-graduation e-mail supplied by the student, we have no easy way
to match it against a NetID or PPID.

This change displays the e-mail associated with the user so that we
can search and sort on e-mail addresses - and therefore match
bouncing addresses to the corresponding student or approver.