Add support for graduating undergraduate dual majors
The undergraduate college and business school have launched a dual
major program that allows students to complete both a liberal arts
and a business undergraduate degree.

**ISSUE**
Emory would like these student to list themselves as being part of
the "Emory College" program, but the Registrar is showing their
graduation record associated with the "Undergraduate Business" school.

This means the application expects to find a graudation record like
`PPIDxxx-UCOL-BBA`, but the Registrar data provides `PPIDxxx-UBUS-BBA`.

There are other students graduating in "UCOL" programs, so we can't
just change the mapping; we have to add an additional check for this
specific case.

**SOLUTION**
This commit adds a secondary check when-and-only-when the student
has entered "Emory College" and a "B.B.A" degree that also scans
Registrar data for the same student and degree within the "Undergraduate
Business" program.