Add job to process registrar data asynchronously
Create a job that wraps the GraduationService and allows it to run
as a background job.  This allows an administrator to upload
registrar data via the dashboard and then have it processed in the
background.  On commpletion, the job updates the RegistrarFeed
status and attached the graduation report so admins can
download it on-demand.