Refactor the GraduationService parameters
The previous interface for passing data between the RegistrarFeed
and GraduationService and back was awkward.

This change lets the GraduationService accept the corresponding
RegistrarFeed and read and update it directly.  This simplifies the
RegistrarJob class and moves all the heavy lifting into the service.

This completes the work necessary to submit registrar data and run
the graduation and publication process entirely through the
Administrator UI.  This is more secure* than running via the console,
so we can also remove the corresponding rake task.

* The process can now only be run by an authenticated user with
super admin priveleges.  And, by uploading data directly to the
application, we're less likely to leave copies floating around in
local download directories or support tickets.