Remove unused RegistrarFeed actions and clean up views
Once registrar data is uploaded, it can't be modified by users,
so we don't need edit and update actions.

All of the relevant information is shown in the index view, so
we don't need a dedicated show view.

The feeds provide audit and provenance data, so we're not allowing
them to be deleted via the UI (any cleanup would need to happen
manually in the console or have a new task/feature written)