Fix failing selection tests
**ISSUE:**
Over time the graduation date authority (which holds a list of
active semesters) gets updated and any specifcic semester ages
out of the list.  Therefore, any tests that rely on specific
entries in the list will eventually break when the corresponding
semester becomes inactive.

**FIX:**
Dynamically look up the value of the first active entry in the
graduation date authority and use it for testing.  This is
slightly less concrete than using string literals, but makes the
tests less fragile.