Implement stubbed out virus scanner for testing
The CI environment does not have virus scanning software installed
and we don't really want to test the behavior of the external
service.

This change gives us an clean way to test the behavior of our app
when the virus scanner flags a file as being infected.

Since the attach_files_to_works job calls the scanner, many tests
run and ignore the scanner.  By stubbing out the rest of the scanner,
we get faster performance in these other tests that don't evaluate
the results of virus scans.