Add junit coverage reporter to yarn dependencies
**ISSUE**
The current test suite installs the junit test reporter on the CI system
each time the test suite is run.  This step alone takes over 1 minute
or approimately 10% of the total test time.

**SOLUTION**
Add the package to the development/test environment so that it gets
cached with other yarn packages rather than having to be installed
fresh on every test run.