Allow GraduationService to process CSV data
The registrar provides data in CSV format and folks have been
converting it to JSON for processing by the application.

Rather than converting the data to JSON and then parsing it back
to a hash from JSON, this change enables the application to
process the CSV file directly from the registrar.

JSON support is still provided for backwards compatibility. This
also allows the change to be deployed to production independently of
cutting over the data format in the registrar workflow.