Update the order of semesters in the submission form
The ETD submission form resets the graduation semester whenever
changes are made to a submission before approval and publication.
A number of times, students and approvers have missed this change
and ETD graduation semesters have been changed inadvertently.

After discussion with SCO, we decided to re-order the list so the
current graduation semester is always at the top of the selection
box.  While this does not address the root problem, we believe that
it will nearly eliminate the issue since most edits are being made
for submissions in the current semester.