Handle CSV uploads from Windows systems
Some Microsoft Windows systems may upload CSV files with a mime
type of 'application/vnd.ms-excel' instead of 'text/csv'.

This change allows the GraduationService to process files with the
microsoft mime-type as a csv.  File upload is still restricted to
CSV by the accept filter on the upload form, and other sanity checks in the
GraduationService will flag if the uploaded file does not appear to contain
valid CSV data.

For more detail on the mime-type issue see
https://christianwood.net/posts/csv-file-upload-validation/