Fix embargo bug in ETD presenter
**ISSUE**
ETDs with expired embargos are displaying abstract, toc, and files
as still being under embargo past the embargo exipration date.

**ANALYSIS**
Rather than checking whether the embargo is active, the code was
only checking for the presence of an embargo release date - even
if that date was in the past.

**RESOLUTION**
Update the presenter to check whether the ETD is under embargo
rather than just checking for the presence of a release date.