[FIX] Graduation dates are reset when editing ETDs
**ISSUE**
When a previously saved ETD is edited, the system would reset the
graduation date to the current semester, regardless of the originally
submitted value.

**DIAGNOSIS**
A bug in the code that reloaded the ETD data only loaded the first
letter of the field. Because this value didn't match any of the
values in the graduation date config file, the form replaced the
unknown value with the current default.

**RESOLUTION**
Fix the bug in the code that reloads saved ETD data into the form,
so that the existing value is preserved and displayed in the edit form.