[REFACTOR] Set coverage reporting default to off
**ISSUE**
Rather than having to always set an environment variable to disable
coverage if we don't want it; change the configuration so that
coverage reporting is only generated during the test suite when it
is explicitly requested, e.g.

```
COVERAGE=true bundle exec rspec
```

Otherwise, run the test suite without generating a coverage report.