[REFACTOR] Update WorkflowSetup to use SemanticLogger
**ISSUE**
WorkflowSetup was using it's own Logger instance and defaulting
to sent logs to STDOUT. This meant logs weren't being persisted
anywhere, making it difficult to debug the class.

**RESOULTION**
Refactor the code to use SemanticLogger like the rest of the
application. We have the application setup to log to the default
log for the Rails environment (e.g. test --> log/test.log,
production --> log/production.log). So WorkflowSetup inherits
the application default and it's easy to find the log for debugging.