[REFACTOR] Make "Emory One-Step Approval" the default workflow
**RATIONALE**
Hyrax ships with a model 'default' workflow which can be used in
simple repository setups. Since the ETD application uses highly
customized workflows, we don't require the default. Having this
additional, unused, workflow adds a small amount of time to
workflow and creates noise in setup logs.

This change removes the default workflow and sets the application
to use the "Emory One-Step Approval" if a workflow is not explicity
set.