Display Woodruff/Nursing as school option on ETD submission form
**ISSUE**
The VUE.js forms have their own configuration data and don't inherit
from the YAML config files in all cases. Even though we added the Nursing
school to the config files and added workflow tests, the school was
not showing up in the list of options in the submission form.

**RESOLUTION**
Update the JavaScript (VUE) setup file to include the Nursing school
as an option along with displaying associated programs in the Program
tab.

**NOTE**
This change also renames the "nursing" proram authority to "woodruff"
for consistency with the way other schools are set up.