[FIX] Department label incorrect when editing Nursing
**ISSUE**
The label for the department/specialty field is displaying
"Department" instead of "Specialty" when editing an InProgressEtd
(i.e. and ETD that has been submitted and is being edited in single
form view instead via the tabbed wizard).

**RESOLUTION**
The VUE Department component has duplicated code that did not inlcude
the dynamic field label call. There were no other differences between
the two sections of code, so we removed the v-if conditional and
removed the duplicated code.