Update deparment selection to be context-sensitive
**ISSUE**
The department selection prompt text displayed the static string
"Select a Department" desptite the input label being dynamic based
on the school. This led to the slightly confusing input
"Specialty: Select a Department" for the school of nursing.

**RESOLUTION**
This change updates the prompt text to match the input label, based
on the previously selected shcool.