[FIX] Embargo selections reset when editing ETDs
**ISSUE**
When a student or admin edits a submitted ETD, the "Content to Embargo"
setting is always reset to "Files Embargoed" regardless of what was
previously saved.

**DIAGNOSIS**
The library (and corresponding tests) that converts saved embargo
settings used a string values instead of boolean values. The source
ETD stores `files_embargoed`, `toc_embargoed`, and `abstract_embargoed`
as boolean values, bt the translation code was comparing against string
values. Examples:
```
true == 'true'
=> false

[true, true, false] == ['true', 'true', 'false']
=> false
```

**RESOLUTION**
Add test cases that demonstrate the bug and update the code to
do the comparisions using the correct data types.