[REFACTOR] Remove redundant embargo code
**ISSUE**
The controller logic to parse embargo codes is only applied on
updates, but not new ETD creation. We didn't notice this because the
same functionality is provided in the actor stack, see:
https://github.com/curationexperts/laevigata/blob/v12.38.4/app/actors/hyrax/actors/etd_actor.rb#L24-L30

**RESOLUTION**
Since the actor stack logic applies in more contexts (background jobs, etc.)
including controller actions, we can safely remove the redundant
logic from the controller.