[REFACTOR] Simplify embargo_type strings
**RATIONALE**
Rather than mirroring the boolean settings in the embargo_type string
passed to and from the ETD forms, we can just pass the highest level
restriction - i.e. `all_restricted` also implied `toc_restricted` and
`files_restricted`.

Effectively, we only need to pass a unique string for each of the three
possible embargo levels:

|`embargo_type`|`files_embargoed`|`toc_embargoed`|`abstract_embargoed`|
|----------------|---------------|---------------|--------------------|
|'files_restricted'|true         | false         | false              |
|'toc_restricted'| true          | true          | false              |
|'all_restricted'| true          | true          | true               |

By using the same strings that are present in the visibility translator,
we set ourselves up to replace the strings with the corresponding constants
in the Rails codebase.