[REFACTOR] Consolidate similar embargo logic
**RATIONALE**
The application had specific code to translate the Rails embargo
settings into a format usable by the JavaScript ETD forms that
was very similar to existing logic in the VisibilityTranslator class.

Having aligned terminology in the JavaScript and Rails portions of
embargo handling code in previous refactorings, we can now use the
logic from `VilibilityTranlator` class directly and eliminate the
need for the `EmbargoTypeFromAttributes` class.

**NOTE**
This change also significantly speeds up these tests (20s --> 2s)
by replacing FactoryBot `.create` calls with `.build` calls
since none of the tests require the objects to be stored to the
database or Fedora.