[REFACTOR] Remove dead code for no_embargoes
**RATIONALE**
A previous version of the application used a Rails form object instead
of the current VUE.js form for ETD creation and editing. This form
used an attribute called `no_embargo` to indicate that no embargo
had been requested (i.e. open access immediately).

The current codebase relies on the `embargo_length` value to determine
whether to apply an embargo or not, so the legacy form attribute and
related code can be removed.