[REFACTOR] Infer affilitationType from affiliation
**RATIONALE**
The submission form uses `affiliationType` to drive the UI, but only
the `affiliation` is stored permanently. We can always infer the
`affiliationType` from the `affiliation` based on the following rules:

|`affiliation`|`affiliationType`|
|-------------|-----------------|
|*blank/empty*|Emory University |
|Emory University|Emory University|
|*any other value*|Non-Emory|

Moving the logic to infer the type into the Vue application sets us
up to remove it from the Rails application code.