[REFACTOR] Remove unused UI partials
**RATIONALE**
The application currently uses a Vue.js application to provide
the UI for submitting and editing ETDs (via the InProgressEtd model).

This bypasses the Hyrax generated forms, related classes, and views,
which were not removed from the codebase when we cut-over
to the JavaScript based front-end.

Therefore, these Hyrax field-editing partials are no longer
called in the current codebase.