[REFACTOR] Remove unused UI helpers
**RATIONALE**
The application currently uses a Vue.js application to provide the
UI for submitting and editing ETDs (via the InProgressEtd model).

This bypasses the Hyrax generated forms, related classes, and views,
which were not removed from the codebase when we cut-over to
the JavaScript based front-end.

These helpers were only used in editing partials that are not called
in the current UI.