[REFACTOR] Remove dead code - do_not_send list
**ISSUE**
The EmailIntercept class is not currently fully covered by the test
suite. In particular, the config file loading safety check is not
tested. However, we are not actively maintaining this list and prefer
deactivating accounts instead. The only remaining function of this
code is to prevent bounces from the dummy account.

**RESOLUTION**
Eliminate the configuration file since it is no longer in use.
Update the documentation appropriately.