[REFACTOR] Simplify authority service setup
**RATIONALE**
The authority services contained a sinificant amount of duplicated
code with only one initialization term differing in each class.

There was also configuration duplicated between the JavaScript front-end
and services in the Rails application.

This change refactors the services to reduce duplicated code and
to replace mapping logic with existing configuration files.

**NOTE**
This refactoring also allows us to display School of Nursing information
in the school dashboard witout having to add custom Woodruff program
mappings in the LaevigataAuthorityService.