[MAINTENANCE] Add tests for OKComputer ETD test
**ISSUE**
OKComputer inlcudes a check for the existence of an ETD to demonstrate
that the persistence layer is correctly configured. However, since
tests run in random order and may leave artifacts in Solr/Fedora,
so there may or may not be ETDs present when the test run.

**RESOLUTION**
This change adds tests that stub the ETD call for both presence and
absence, allowing us to test both check pathways independently of
the state of persistence when the tests are run in randomized order.