[REFACTOR] SupplementalFile metadata processing in ETD Controlleri
**ISSUE**
The controller has logging code related to an issue in the now
obsolete Rails form. See:
https://github.com/curationexperts/laevigata/commit/82452b536fd7ead086b9dfa4333136fa3906d1a9

The Honeybadger notification is no longer relevant with the
JavaScript submission form.

We've also slightly modified the call structure of supporting
calls and made the respective methods private since they should
not be called outside the controller request context.