[FIX] Admin Dashboard test is failing intermittently
**ISSUE**
The Admin Dashboard test is failing intermittently, but not
consistently in local test and CI environments.

**DIAGNOSIS**
The test suite is sometimes proceeding to reload the ETD in the
spec before turbolinks had submitted the form data.

**RESOLUTION**
Waiting for content that appears after the page refresh ensures
that the JavaScript on the page has had time to finish executing
and the application has recieved the form submission before
proceeding with additional test steps.